TypeScriptified `normalizeEmail`
